Unify paddings on collection page
