+++ src/domain/collection/eventList/eventList.module.scss
@@ -8,4 +8,10 @@
   h2 {
     text-align: center;
   }
+
+  .eventSearchListWrapper {
+    max-width: 68.75rem;
+    margin-left: auto;
+    margin-right: auto;
+  }
 }
