+++ src/domain/eventSearch/searchResultList/eventList.module.scss
@@ -1,0 +1,11 @@
+@import "layout";
+
+.eventList {
+  .loadMoreWrapper {
+    margin-top: 1.5rem;
+
+    @include respond_above(sm) {
+      margin-top: 0;
+    }
+  }
+}
