+++ src/domain/eventSearch/searchResultList/searchResultList.module.scss
@@ -23,19 +23,13 @@
   display: grid;
   grid-row-gap: 0.625rem;
 
+  margin-left: -0.75rem;
+  margin-right: -0.75rem;
+
   @include respond_above(sm) {
     grid-row-gap: unset;
     grid-column-gap: $largeMargin;
-    margin-left: auto;
-    margin-right: auto;
-    max-width: 68.75rem;
-
-    &.hasFilterSummary {
-      grid-template-columns: 2fr 5fr;
-      margin-left: unset;
-      margin-right: unset;
-      max-width: unset;
-    }
+    grid-template-columns: 2fr 5fr;
   }
 
   .eventListWrapper {
